+++ app/src/main/java/woowacourse/movie/notification/NotificationChannelInfo.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui
+package woowacourse.movie.notification
 
 enum class NotificationChannelInfo(val id: String, val description: String) {
     BOOKING_ALARM("영화 예매 알림", "예매한 영화를 30분 전에 알려주는 알림"),
