+++ app/src/main/java/woowacourse/movie/notification/PendingIntentGenerator.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui
+package woowacourse.movie.notification
 
 import android.app.PendingIntent
 import android.content.Context
